requirements.txt
# Requiere Python 3.12.x
chromadb>=0.4.24
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-openai>=0.0.1
langchain-chroma>=0.0.1
openai>=1.0.0
youtube-transcript-api>=0.4.4
pytube>=15.0.0
pandas>=2.0.0
requests>=2.28.0
yt-dlp>=2024.1.1
streamlit>=1.30.0
plotly>=5.18.0
matplotlib>=3.7.0
scipy>=1.10.0
huggingface-hub>=0.20.0
sentence-transformers>=2.2.2
reportlab>=4.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11
ebooklib>=0.18
openai-whisper>=20231117
ffmpeg-python>=0.2.0
torch>=2.0.0
resemblyzer
librosa
soundfile
scikit-learn
python-dotenv
numpy
